Renamed sonar layer to RangeSensorLayer
Since it nicely covers both sonars and IR sensors
