+++ range_sensor_layer/costmap_plugins.xml
@@ -1,0 +1,5 @@
+<library path="lib/librange_sensor_layer">
+  <class type="range_sensor_layer::RangeSensorLayer" base_class_type="costmap_2d::Layer">
+    <description>A range-sensor (sonar, IR) based obstacle layer for costmap_2d</description>
+  </class>
+</library>
